# docs/openapi.yaml
...
openapi: 3.0.3
info:
  title: 'MSP-MIU Website API'
  description: "REST API for the MSP Tech Club MIU platform (members, events, CMS, competitions, quizzes).\n\nCall `POST /auth/login` (or `/users/login`) — the JWT is applied to **Authorize** automatically.\n"
  version: 1.0.0
  contact:
    name: 'MSP Tech Club MIU'
    url: 'https://msp-miu.tech'
servers:
  -
    url: 'https://msp-miu.tech/api'
    description: Production
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
security: []
paths:
...
  /courses:
    get:
      summary: 'Get /courses'
      responses:
        '200':
          description: Success
      tags:
        - courses
    post:
      summary: 'Create/submit /courses'
      responses:
        '200':
          description: Success
      tags:
        - courses
      security:
        -
          bearerAuth: []
  '/courses/{id}':
    get:
      summary: 'Get /courses/{id}'
      responses:
        '200':
          description: Success
      tags:
        - courses
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: string
    put:
      summary: 'Update /courses/{id}'
      responses:
        '200':
          description: Success
      tags:
        - courses
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: string
      security:
        -
          bearerAuth: []
    delete:
      summary: 'Delete /courses/{id}'
      responses:
        '200':
          description: Success
      tags:
        - courses
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: string
      security:
        -
          bearerAuth: []
...